requirements.txt
Flask
tensorflow
keras
Pillow
numpy
opencv-python